requirements.txt
asgiref==3.5.2
astroid==2.11.7
dill==0.3.5.1
Django==3.2.14
django-filter==22.1
django-restframework==0.0.1
djangorestframework==3.13.1
importlib-metadata==4.12.0
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.4.1
mccabe==0.7.0
mysql-connector-python==8.0.30
networkx==2.6.3
numpy==1.21.6
pandas==1.1.5
platformdirs==2.5.2
protobuf==3.20.1
PyHamcrest==2.0.3
pylint==2.13.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.4
typing-extensions==4.3.0
wrapt==1.14.1
zipp==3.8.1
